+++ db/migrate/20181102122138_change_dedicated_for_to_array.rb
@@ -1,9 +1,9 @@
 class ChangeDedicatedForToArray < ActiveRecord::Migration[5.2]
   def up
-    change_column :services, :dedicated_for, :text, using: "(string_to_array(dedicated_for, ','))", array: true
+    change_column :services, :dedicated_for, :string, using: "(string_to_array(dedicated_for, ','))", array: true
   end
 
   def down
-    change_column :services, :dedicated_for, :text, null: false
+    change_column :services, :dedicated_for, :string, null: true
   end
 end